% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endemicchannel.R
\name{endemicchannel}
\alias{endemicchannel}
\title{Create endemic channel}
\usage{
endemicchannel(data, yearnow, period = "month", .var = "cases", nyears = 5)
}
\arguments{
\item{data}{A data frame with columns: year, month/week, cases.}

\item{yearnow}{The year for which the endemic channel will be computed}

\item{period}{Specify the name of the column to aggregate the time: i.e. "month" or "epiweek". Default period="month"}

\item{.var}{Column name to compute the endemic channel. Default .var="cases"}

\item{nyears}{Number of years to compute the endemic channel. Default nyears= 5}
}
\value{
Returns a dataframe with colums: year, month, median, supramin, inframax
}
\description{
Create endemic channel
}
\examples{
set.seed(1)
x=seq(0,12*6-1)
y=100 + 20*cos(2*pi*x/12) + rnorm(n = length(x),mean = 50,sd = 40)
y[y<0] <- 0
data14a19 <- data.frame( idx=seq(1,12*6), year=rep(2014:2019, each=12 ),
 month=rep(1:12, 6), cases=y )
endemicchannel(data =data14a19,yearnow = 2019 )
}
\references{
Baldoquin-Rodríguez W, et al. Effectiveness of a multicomponent dengue prevention strategy targeting transmission hotspots in Santiago de Cuba, Cuba, 2015-2020. 2024. [in-review]
}
\author{
W. Baldoquin Rodríguez
}
